Update windows.md
Hi, I'm a LeWagon student (Melbourne - batch 857)
Just wanted to try contributing on GitHub 😄 
I've added number emojis to level 2 headings and arrow emojis to level three headings. Just wanted to make the headings standout a bit more as I felt they got lost in all the text. This made it a little hard for me to skim read / identify which step I was up to when asked. 
Hope this helps! 🚀 
Thanks